classify/requirements.txt
numpy
pillow
pdf2image
torch
torchvision
flask